added rooms and abstracted away the server code
