+++ lib/io.js
@@ -30,9 +30,11 @@ rtc.initializedStreams = 0;
 /**
  * Connects to the socket.io server.
  */
-rtc.connect = function(server) {
+rtc.connect = function(server, room) {
+  room = room || ''; // by default, join a room called the blank string
   rtc._socket = io.connect(server);
   rtc._socket.on('connect', function() {
+    rtc._socket.emit('join room', room);
     rtc.fire('connect');
   });
